invoice: swap PaymentSecret for ChannelManager's PaymentSecret
